feat: add internationalization functionality
