+++ locale/en/index.js
@@ -1,0 +1,11 @@
+const getActiveWords = require('./getActiveWords');
+const gridColumns = require('./gridColumns');
+const letters = require('./letters');
+const wordMap = require('./wordMap');
+
+module.exports = {
+  getActiveWords,
+  gridColumns,
+  letters,
+  wordMap,
+};
