+++ locale/en/letters.js
@@ -1,0 +1,112 @@
+module.exports = [
+  'i',
+  't',
+  'l',
+  'i',
+  's',
+  'a',
+  's',
+  'a',
+  'm',
+  'p',
+  'm',
+  'a',
+  'c',
+  'q',
+  'u',
+  'a',
+  'r',
+  't',
+  'e',
+  'r',
+  'd',
+  'c',
+  't',
+  'w',
+  'e',
+  'n',
+  't',
+  'y',
+  '-',
+  'f',
+  'i',
+  'v',
+  'e',
+  'h',
+  'a',
+  'l',
+  'f',
+  's',
+  't',
+  'e',
+  'n',
+  'f',
+  't',
+  'o',
+  'p',
+  'a',
+  's',
+  't',
+  'e',
+  'r',
+  'u',
+  'n',
+  'i',
+  'n',
+  'e',
+  'o',
+  'n',
+  'e',
+  's',
+  'i',
+  'x',
+  't',
+  'h',
+  'r',
+  'e',
+  'e',
+  'f',
+  'o',
+  'u',
+  'r',
+  'f',
+  'i',
+  'v',
+  'e',
+  't',
+  'w',
+  'o',
+  'e',
+  'i',
+  'g',
+  'h',
+  't',
+  'e',
+  'l',
+  'e',
+  'v',
+  'e',
+  'n',
+  's',
+  'e',
+  'v',
+  'e',
+  'n',
+  't',
+  'w',
+  'e',
+  'l',
+  'v',
+  'e',
+  't',
+  'e',
+  'n',
+  's',
+  'e',
+  "o'",
+  'c',
+  'l',
+  'o',
+  'c',
+  'k',
+];
