+++ locale/en/wordMap.js
@@ -1,0 +1,26 @@
+module.exports = {
+  it: [0, 1],
+  is: [3, 4],
+  a: [11],
+  quarter: [13, 14, 15, 16, 17, 18, 19],
+  twenty: [22, 23, 24, 25, 26, 27],
+  five: [29, 30, 31, 32],
+  twentyfive: [22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32],
+  half: [33, 34, 35, 36],
+  ten: [38, 39, 40],
+  past: [44, 45, 46, 47],
+  to: [42, 43],
+  1: [55, 56, 57],
+  2: [74, 75, 76],
+  3: [61, 62, 63, 64, 65],
+  4: [66, 67, 68, 69],
+  5: [70, 71, 72, 73],
+  6: [58, 59, 60],
+  7: [88, 89, 90, 91, 92],
+  8: [77, 78, 79, 80, 81],
+  9: [51, 52, 53, 54],
+  10: [99, 100, 101],
+  11: [82, 83, 84, 85, 86, 87],
+  12: [93, 94, 95, 96, 97, 98],
+  oclock: [104, 105, 106, 107, 108, 109],
+};
