+++ locale/index.js
@@ -1,0 +1,5 @@
+const en = require('./en');
+
+module.exports = {
+  en,
+};
