+++ node_helper.js
@@ -1,0 +1,23 @@
+const NodeHelper = require('node_helper');
+const locale = require('./locale');
+
+module.exports = NodeHelper.create({
+  socketNotificationReceived: function (notification, payload) {
+    if (notification === 'SET_LANGUAGE') {
+      this.sendSocketNotification(
+        'SET_LANGUAGE',
+        JSON.stringify(
+          locale[payload],
+          (_, value) => {
+            if (typeof value == 'function') {
+              return '__FUNC__' + value.toString();
+            }
+
+            return value;
+          },
+          2
+        )
+      );
+    }
+  },
+});
